LPD-58844 Move getMailId logic from Portal to MailService and apply
