+++ modules/apps/mail/mail-messaging-impl/src/main/java/com/liferay/mail/messaging/internal/MailServiceImpl.java
@@ -30,6 +30,7 @@ import com.liferay.portal.kernel.util.PropertiesUtil;
 import com.liferay.portal.kernel.util.PropsKeys;
 import com.liferay.portal.kernel.util.PropsUtil;
 import com.liferay.portal.kernel.util.Validator;
+import com.liferay.portal.util.PropsValues;
 
 import jakarta.mail.Authenticator;
 import jakarta.mail.PasswordAuthentication;
@@ -76,6 +77,40 @@ public class MailServiceImpl
 		return new Class<?>[] {
 			MailService.class, IdentifiableOSGiService.class
 		};
+	}
+
+	@Override
+	public String getMailId(String mx, String popPortletPrefix, Object... ids) {
+		StringBundler sb = new StringBundler((ids.length * 2) + 7);
+
+		sb.append(StringPool.LESS_THAN);
+		sb.append(popPortletPrefix);
+
+		if (!popPortletPrefix.endsWith(StringPool.PERIOD)) {
+			sb.append(StringPool.PERIOD);
+		}
+
+		for (int i = 0; i < ids.length; i++) {
+			Object id = ids[i];
+
+			if (i != 0) {
+				sb.append(StringPool.PERIOD);
+			}
+
+			sb.append(id);
+		}
+
+		sb.append(StringPool.AT);
+
+		if (Validator.isNotNull(PropsValues.POP_SERVER_SUBDOMAIN)) {
+			sb.append(PropsValues.POP_SERVER_SUBDOMAIN);
+			sb.append(StringPool.PERIOD);
+		}
+
+		sb.append(mx);
+		sb.append(StringPool.GREATER_THAN);
+
+		return sb.toString();
 	}
 
 	@Override
