+++ modules/dxp/apps/multi-factor-authentication/multi-factor-authentication-email-otp-web/src/main/java/com/liferay/multi/factor/authentication/email/otp/web/internal/portlet/action/SendMFAEmailOTPMVCResourceCommand.java
@@ -105,7 +105,7 @@ public class SendMFAEmailOTPMVCResourceCommand implements MVCResourceCommand {
 			toUser.getCompanyId());
 
 		mailMessage.setMessageId(
-			_portal.getMailId(
+			_mailService.getMailId(
 				company.getMx(), "user", toUser.getUserId(),
 				System.currentTimeMillis()));
